LayoutTests/inspector/dom-debugger/event-timeout-breakpoints-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for Event Timer breakpoints.


== Running test suite: DOMDebugger.Event.Timer
-- Running test case: DOMDebugger.Event.Timeout.AddBreakpoint
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleWindow_setTimeout
ASYNC CALL STACK:
1: --- setTimeout ---
2: [F] trigger_setTimeout
3: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Timeout.AddDisabledBreakpoint
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...
Disabling "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Timeout.RemoveBreakpoint
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...
Removing "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Timeout.RemoveDisabledBreakpoint
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...
Disabling "timeout" Event Breakpoint...
Removing "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Timer.Options.Condition
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Timer.Options.IgnoreCount
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Timer.Options.Action.Log
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Timer.Options.Actions.Evaluate
Creating "timeout" Event Breakpoint...
Adding "timeout" Event Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
